pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "container-registry-cli"
version = "1.0.0"
description = "Container image registry analyzer with cleanup policies and vulnerability scanning"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
dependencies = [
    "click>=8.0.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
]

[project.scripts]
container-registry-cli = "container_registry_cli.cli:main"

[tool.setuptools.packages.find]
include = ["container_registry_cli*"]
